api_gateway_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
redis[asyncio]==5.2.0
python-multipart==0.0.12